refactor: 참가자(Participant) 추상 클래스 구현
